Basic project structure with unit tests
