--- src/attention_broker.cc
@@ -1,6 +1,0 @@
-#include "attention_broker.h"
-
-Status AttentionBrokerServer::ping(ServerContext* context, const das::Empty* request, das::Ack* reply) {
-    reply->set_msg("Hey");
-    return Status::OK;
-}
