+++ src/grpc_server/BUILD
@@ -1,0 +1,13 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "grpc_server_lib",
+    srcs = glob(["*.cc"]),
+    hdrs = glob(["*.h"]),
+    includes = ["."],
+    deps = [
+        "@com_github_singnet_das_proto//:attention_broker_cc_grpc",
+        "@com_github_grpc_grpc//:grpc++",
+        "@com_github_grpc_grpc//:grpc++_reflection",
+    ],
+)
