+++ src/grpc_server/attention_broker_server.cc
@@ -1,0 +1,12 @@
+#include "attention_broker_server.h"
+
+AttentionBrokerServer::AttentionBrokerServer() {
+}
+
+AttentionBrokerServer::~AttentionBrokerServer() {
+}
+
+Status AttentionBrokerServer::ping(ServerContext* context, const das::Empty* request, das::Ack* reply) {
+    reply->set_msg("OK");
+    return Status::OK;
+}
