+++ src/tests/BUILD
@@ -8,7 +8,10 @@ cc_test(
     ],
     deps = [
         "@com_github_google_googletest//:gtest_main",
-        "//:attention_broker",
+        "@com_github_singnet_das_proto//:attention_broker_cc_grpc",
+        "@com_github_grpc_grpc//:grpc++",
+        "@com_github_grpc_grpc//:grpc++_reflection",
+        "//grpc_server:grpc_server_lib",
     ],
 )
